+++ src/main/java/qna/domain/Question.java
@@ -3,9 +3,7 @@ package qna.domain;
 import qna.CannotDeleteException;
 
 import javax.persistence.*;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Entity
 @Table(name = "question")
@@ -24,8 +22,8 @@ public class Question extends BaseEntity {
 	@JoinColumn(name = "writer_id", foreignKey = @ForeignKey(name = "fk_question_writer"))
 	private User writer;
 
-	@OneToMany(mappedBy = "question", fetch = FetchType.LAZY)
-	private final List<Answer> answers = new ArrayList<>();
+	@Embedded
+	private Answers answers = new Answers();
 
 	@Column(nullable = false)
 	private boolean deleted = false;
@@ -52,10 +50,10 @@ public class Question extends BaseEntity {
 	}
 
 	public void addAnswerData(Answer answer) {
-		this.answers.add(answer);
+		answers.addAnswer(answer);
 	}
 
-	public List<Answer> getAnswers() {
+	public Answers getAnswers() {
 		return answers;
 	}
 
@@ -77,9 +75,7 @@ public class Question extends BaseEntity {
 
 	public DeleteHistories delete(User loggedInUser) throws CannotDeleteException {
 		validateWriter(loggedInUser);
-		validateAnswersOwner(loggedInUser);
-		List<DeleteHistory> deleteHistories = this.answers.stream().map(Answer::delete)
-				.collect(Collectors.toList());
+		List<DeleteHistory> deleteHistories = answers.deleteAnswers(loggedInUser);
 		this.deleted = true;
 		deleteHistories.add(new DeleteHistory(ContentType.QUESTION, id, loggedInUser));
 		return new DeleteHistories(deleteHistories);
@@ -89,11 +85,6 @@ public class Question extends BaseEntity {
 		if (!this.writer.matchUser(loggedInUser)) {
 			throw new CannotDeleteException("질문을 삭제할 권한이 없습니다.");
 		}
-	}
-
-	private void validateAnswersOwner(User loggedInUser) throws CannotDeleteException {
-		Answers answers = new Answers(this.answers);
-		answers.validateAnswersOwner(loggedInUser);
 	}
 
 	@Override
